// src/types/interfaces/camera.interface.d.ts
import mongoose, { Schema } from "mongoose";
export { };

//define camera type
declare global {
  interface ICamera extends Document {
    _id: mongoose.Types.ObjectId;
    section_id: mongoose.Types.ObjectId;
    network: string;
    url: string;
    nvr: string;
    ip: string;
    name: string;
    username: string;
    password: string;
    muted: Schema.Types.ObjectId[];
    is_enabled: boolean;
    create_date: Date;
    camera_type: string;
  };

  interface ICameraInfo {
    [key: string]: string
    ip: string;
    username: string;
    password: string;
    nvr: string;
  };

};